Added full restful to every routes + updated old routes and doc

+++ routes/restaurants/products.js
@@ -4,7 +4,6 @@ const mongoose = require('mongoose');
 
 const { Schema } = mongoose;
 const productSchema = new Schema({
-    IdProduct : Number,
     IdRestaurant: Number,
     Name: String,
     Description: String,
@@ -41,7 +40,7 @@ const Product = mongoose.model('Product', productSchema);
  * 
  * @apiDescription Returns an array of information.
  * 
- * @apiSuccess {Number} IdProduct  Product's unique id.
+ * @apiSuccess {String} _id  Product's unique id.
  * @apiSuccess {Number} IdRestaurant  Restaurants's id related to this product.
  * @apiSuccess {String} Name  Name of this product.
  * @apiSuccess {String} Description  Description of this product.
@@ -55,7 +54,7 @@ const Product = mongoose.model('Product', productSchema);
 router.get('/', function(req, res) 
 {
     Product.find({
-        id_restau: JSON.parse(req.params.idRestaurant)
+        IdRestaurant: JSON.parse(req.params.IdRestaurant)
     }, (err, docs) => {
         if (err)
             res.status(500).json({ message: "ProductsNotAccessible" });
@@ -90,9 +89,9 @@ router.get('/', function(req, res)
  * @apiName GetProduct
  * @apiGroup Products
  * 
- * @apiParam {Number} id  Product's unique id.
+ * @apiParam {String} IdProduct  Product's unique id.
  * 
- * @apiSuccess {Number} IdProduct  Product's unique id.
+ * @apiSuccess {String} _id  Product's unique id.
  * @apiSuccess {Number} IdRestaurant  Restaurants's id related to this product.
  * @apiSuccess {String} Name  Name of this product.
  * @apiSuccess {String} Description  Description of this product.
@@ -103,13 +102,11 @@ router.get('/', function(req, res)
  *
  * @apiError ProductNotFound The products was not found.
  */
-router.get('/:id', function(req, res) 
+router.get('/:IdProduct', function(req, res) 
 {
     Product.findOne({
-        where: {
-            id_restau: JSON.parse(req.params.idRestaurant),
-            id: JSON.parse(req.params.id)
-        }
+        IdRestaurant: req.params.IdRestaurant,
+        _id: req.params.IdProduct
     }).then(function(product) {
         if (product) {
             return res.status(200).json(product);
@@ -163,7 +160,7 @@ router.post('/', function(req, res)
 {
     Product.findOne({
         where:  {
-            IdRestaurant: req.body.IdRestaurant,
+            IdRestaurant: req.params.IdRestaurant,
             Name: req.body.Name
         }
     }).then(function(product) {
@@ -174,7 +171,7 @@ router.post('/', function(req, res)
         }
 
         Product.create({
-            IdRestaurant: req.body.IdRestaurant,
+            IdRestaurant: req.params.IdRestaurant,
             Name: req.body.Name,
             Description: req.body.Description,
             Picture: req.body.Picture,
@@ -225,7 +222,7 @@ router.post('/', function(req, res)
  * @apiName PutProducts
  * @apiGroup Products
  * 
- * @apiParam {Number} id  Product's unique id.
+ * @apiParam {String} IdProduct  Product's unique id.
  * @apiParam {Number} IdRestaurant  Restaurants's id related to this product.
  * @apiParam {String} Name  Name of this product.
  * @apiParam {String} Description  Description of this product.
@@ -237,20 +234,41 @@ router.post('/', function(req, res)
  * @apiSuccess {String} message Products updated.
  *
  * @apiError ProductNotUpdated The product cannot be updated.
+ * @apiError ProductNotExisting The product does not exists.
+ * @apiError DatabaseError Database issues.
  */
-router.put('/:id', function(req, res) 
+router.put('/:IdProduct', function(req, res) 
 {
-    Product.updateOne({
-        id_restau: JSON.parse(req.params.idRestaurant),
-        id: JSON.parse(req.params.id)
-    }, 
-    req.body,
-    function (err, docs) 
-    {
-      if (err)
-        res.status(401).json({ message: "ProductNotUpdated" });
-      else
-        res.status(202).json({ message: "Products updated" });
+    Product.findById(req.params.IdProduct).then(function(product) {
+        if (!product) {
+            return res.status(403).json({
+                message: 'ProductNotExisting'
+            });
+        }
+
+        Product.updateOne({
+            IdRestaurant: req.params.IdRestaurant,
+            Name: req.body.Name,
+            Description: req.body.Description,
+            Picture: req.body.Picture,
+            Size: req.body.Size,
+            Notes: req.body.Notes,
+            VoteNb: req.body.VoteNb
+        }).then((response) => {
+            return res.status(201).json({
+                message: 'Product updated'
+            });
+        }).catch((error) => {
+            return res.status(401).json({
+                message: 'ProductNotUpdated',
+                stackTrace: error
+            });
+        });
+    }).catch((error) => {
+        return res.status(500).json({
+            message: 'DatabaseError',
+            stackTrace: error
+        })
     });
 });
 
@@ -273,23 +291,41 @@ router.put('/:id', function(req, res)
  * @apiName DeleteProducts
  * @apiGroup Products
  * 
- * @apiParam {Number} id  Product's unique id.
+ * @apiParam {Number} IdProduct  Product's unique id.
  * 
  * @apiSuccess {String} message Products deleted.
  *
  * @apiError ProductNotDeleted The product cannot be deleted.
+ * @apiError ProductNotExisting The product cannot be found.
+ * @apiError DatabaseError Database issues.
  */
-router.delete('/:id', function(req, res)
+router.delete('/:IdProduct', function(req, res)
 {
-    Product.deleteOne({
-        id_restau: JSON.parse(req.params.idRestaurant),
-        id: JSON.parse(req.params.id)
-    }, function (err, docs) 
-    {
-      if (err)
-        res.status(401).json({ message: "ProductNotDeleted" });
-      else
-        res.status(203).json({ message: "Products deleted" });
+    Product.findById(req.params.IdProduct).then(function(product) {
+        if (!product) {
+            return res.status(403).json({
+                message: 'ProductNotExisting'
+            });
+        }
+
+        Product.deleteOne({
+            _id: req.params.IdProduct,
+            IdRestaurant: req.params.IdRestaurant,
+        }).then((response) => {
+            return res.status(201).json({
+                message: 'Product deleted'
+            });
+        }).catch((error) => {
+            return res.status(401).json({
+                message: 'ProductNotDeleted',
+                stackTrace: error
+            });
+        });
+    }).catch((error) => {
+        return res.status(500).json({
+            message: 'DatabaseError',
+            stackTrace: error
+        })
     });
 });
